requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
pydantic==2.6.2
numpy>=1.26,<1.28
pandas==2.1.1

# PyTorch for model inference (CPU-only, smaller footprint)
torch
torchvision

# Optional: HTTP requests and environment variables
requests==2.32.1
python-dotenv==1.0.1

scikit-learn
gpt4all
bertopic
faiss-cpu
